requirements.txt
# Crevia Analytics - Full dependency list
# Generated from pip freeze (Python 3.13, Windows dev environment)

# Core
requests==2.32.5
python-dotenv==1.2.1
pandas==3.0.0
numpy==2.4.1
scipy==1.17.0

# Claude AI
anthropic>=0.77.0

# Web scraping / RSS
beautifulsoup4>=4.12.0
feedparser==6.0.12

# Browser automation (X/Substack posting)
playwright==1.58.0

# Social / APIs
tweepy>=4.14.0
praw>=7.8.1
notion-client>=3.0.0

# FastAPI backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
pydantic>=2.0.0
python-multipart>=0.0.6
httpx>=0.24.0

# Auth
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# Crypto / Base L2 payment verification
web3>=7.0.0

# Utilities
python-dateutil>=2.9.0
pytz>=2025.0
dateparser>=1.2.0
python-docx>=1.2.0
